+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/BinaryDataExporter.java
@@ -24,6 +24,7 @@ import com.jpexs.decompiler.flash.configuration.Configuration;
 import com.jpexs.decompiler.flash.exporters.settings.BinaryDataExportSettings;
 import com.jpexs.decompiler.flash.tags.DefineBinaryDataTag;
 import com.jpexs.decompiler.flash.tags.Tag;
+import com.jpexs.decompiler.flash.tags.base.BinaryDataInterface;
 import com.jpexs.helpers.Helper;
 import com.jpexs.helpers.Path;
 import java.io.BufferedOutputStream;
@@ -44,69 +45,73 @@ import java.util.Set;
 public class BinaryDataExporter {
 
     public List<File> exportBinaryData(AbortRetryIgnoreHandler handler, String outdir, ReadOnlyTagList tags, BinaryDataExportSettings settings, EventListener evl) throws IOException, InterruptedException {
+        List<BinaryDataInterface> binaryDatas = new ArrayList<>();
+        for (Tag t : tags) {
+            if (t instanceof BinaryDataInterface) {
+                binaryDatas.add((BinaryDataInterface) t);
+            }
+        }
+        return exportBinaryData(handler, outdir, binaryDatas, settings, evl);
+    }
+    
+    public List<File> exportBinaryData(AbortRetryIgnoreHandler handler, String outdir, List<BinaryDataInterface> binaryDatas, BinaryDataExportSettings settings, EventListener evl) throws IOException, InterruptedException {
         List<File> ret = new ArrayList<>();
         if (Thread.currentThread().isInterrupted()) {
             return ret;
         }
 
-        if (tags.isEmpty()) {
+        if (binaryDatas.isEmpty()) {
             return ret;
         }
 
         File foutdir = new File(outdir);
         Path.createDirectorySafe(foutdir);
 
-        int count = 0;
-        for (Tag t : tags) {
-            if (t instanceof DefineBinaryDataTag) {
-                count++;
-            }
-        }
+        int count = binaryDatas.size();
 
         if (count == 0) {
             return ret;
         }
 
         int currentIndex = 1;
-        for (final Tag t : tags) {
-            if (t instanceof DefineBinaryDataTag) {
-                DefineBinaryDataTag bdt = (DefineBinaryDataTag) t;
-                if (evl != null) {
-                    evl.handleExportingEvent("binarydata", currentIndex, count, t.getName());
-                }
-
-                String ext = bdt.innerSwf == null ? ".bin" : ".swf";
-                final File file = new File(outdir + File.separator + Helper.makeFileName(bdt.getCharacterExportFileName() + ext));
-                new RetryTask(() -> {
-                    try (OutputStream fos = new BufferedOutputStream(new FileOutputStream(file))) {
-                        fos.write(bdt.binaryData.getRangeData());
-                    }
-                }, handler).run();
-                
-                Set<String> classNames = bdt.getClassNames();                    
-                if (Configuration.as3ExportNamesUseClassNamesOnly.get() && !classNames.isEmpty()) {
-                    for (String className : classNames) {
-                        File classFile = new File(outdir + File.separator + Helper.makeFileName(className + ext));
-                        new RetryTask(() -> {
-                            Files.copy(file.toPath(), classFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
-                        }, handler).run();
-                        ret.add(classFile);
-                    }
-                    file.delete();
-                } else {                
-                    ret.add(file);
-                }
-
-                if (Thread.currentThread().isInterrupted()) {
-                    break;
-                }
-
-                if (evl != null) {
-                    evl.handleExportedEvent("binarydata", currentIndex, count, t.getName());
-                }
-
-                currentIndex++;
+        for (final BinaryDataInterface t : binaryDatas) {
+            if (evl != null) {               
+                evl.handleExportingEvent("binarydata", currentIndex, count, t.getName());                
             }
+
+            String ext = t.getInnerSwf() == null ? ".bin" : ".swf";
+            final File file = new File(outdir + File.separator + Helper.makeFileName(t.getCharacterExportFileName() + ext));
+            new RetryTask(() -> {
+                try (OutputStream fos = new BufferedOutputStream(new FileOutputStream(file))) {
+                    fos.write(t.getDataBytes().getRangeData());
+                }
+            }, handler).run();
+            
+            DefineBinaryDataTag bdt = (DefineBinaryDataTag) t.getTopLevelBinaryData();
+
+            Set<String> classNames = bdt.getClassNames();                    
+            if (Configuration.as3ExportNamesUseClassNamesOnly.get() && !classNames.isEmpty()) {
+                for (String className : classNames) {
+                    File classFile = new File(outdir + File.separator + Helper.makeFileName(t.getClassExportFileName(className) + ext));
+                    new RetryTask(() -> {
+                        Files.copy(file.toPath(), classFile.toPath(), StandardCopyOption.REPLACE_EXISTING);
+                    }, handler).run();
+                    ret.add(classFile);
+                }
+                file.delete();
+            } else {                
+                ret.add(file);
+            }
+
+            if (Thread.currentThread().isInterrupted()) {
+                break;
+            }
+
+            if (evl != null) {
+                evl.handleExportedEvent("binarydata", currentIndex, count, t.getName());
+            }
+
+            currentIndex++;            
         }
 
         return ret;